requirements.txt
anyio==4.1.0
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
autopep8==1.5.5
Babel==2.14.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
black==23.11.0
bleach==6.1.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.0.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
comm==0.2.0
command-not-found==0.3
cryptography==3.4.6
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.2.6
django-bootstrap-v5==1.0.11
django-bootstrap4==23.4
django-crispy-forms==2.1
django-phonenumber-field==7.2.0
django-pytest==0.2.0
djangorestframework==3.14.0
duplicity==0.8.21
exceptiongroup==1.2.0
executing==2.0.1
fasteners==0.14.1
fastjsonschema==2.19.0
filelock==3.13.1
fqdn==1.5.1
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
greenlet==1.1.2
gyp==0.1
httplib2==0.20.2
idna==2.10
importlib-metadata==4.6.4
iniconfig==2.0.0
ipykernel==6.27.1
ipython==8.18.1
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
ndg-httpsclient==0.5.1
nest-asyncio==1.5.8
netifaces==0.11.0
notebook_shim==0.2.3
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
overrides==7.4.0
packaging==23.2
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
phonenumbers==8.13.27
Pillow==10.1.0
platformdirs==3.11.0
pluggy==1.3.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==3.12.4
psutil==5.9.6
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pusher==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycodestyle==2.6.0
pycparser==2.20
pycups==2.0.1
PyDrive==1.3.1
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyRFC3339==1.1
pytest==8.0.0
pytest-django==4.7.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.15.0
python-json-logger==2.0.7
pytz==2023.3.post1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.2
referencing==0.32.0
reportlab==3.6.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.2
rsa==4.8
SecretStorage==3.3.1
Send2Trash==1.8.2
six==1.15.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.31
sqlparse==0.4.4
stack-data==0.6.3
systemd-python==234
terminado==0.18.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.4
traitlets==5.14.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
uritemplate==3.0.1
urllib3==1.26.3
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.24.6
wadllib==1.3.6
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
xdg==5
xkit==0.0.0
zipp==1.0.0
